Final bits of documentation

+++ tests/Quarer.Tests/DataSegmentTests.cs
@@ -18,7 +18,7 @@ public class DataSegmentTests
         var dataLength = Random.Shared.Next(1, 100);
         var version = QrVersion.GetVersion(1);
         var characterCount = CharacterCount.GetCharacterCountBitCount(version, ModeIndicator.Numeric);
-        var segment = DataSegment.Create(characterCount, ModeIndicator.Numeric, dataLength, .., new EciCode(26));
+        var segment = DataSegment.Create(characterCount, ModeIndicator.Byte, dataLength, .., new EciCode(26));
         Assert.Equal(dataLength + 4 + 4 + 8 + characterCount, segment.FullSegmentLength);
     }
